requirements.txt
alembic==1.0.2
Click==7.0
Flask==1.0.2
Flask-Migrate==2.3.0
Flask-SQLAlchemy==2.3.2
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
pkg-resources==0.0.0
pyserial==3.4
python-dateutil==2.7.5
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.2.13
Werkzeug==0.14.1
